requirements.txt
Django==4.2
djangorestframework==3.14.0
celery==5.3.4
psycopg2-binary==2.9.9
django-environ==0.11.2
django-celery-results==2.5.1
django-celery-beat==2.5.0
redis==5.0.1
django-redis==5.4.0
django-filter==23.3
